Changed the random colour function
Also, included a new function randomVal which allows randomization for the randomColor functions to follow.

Can now enter in different values based on HSL - for H - where specific colour values can now be targeted
So, for the first randomVal(10, 11) would be targeting a red to slight orange colour based on those two values. They're also
affected by the S - saturation and the L - lightness which gives a lot of variations on what is a very isolated section of
the colour spectrum.